Fix export contribution to share tasktrait
